[skip ci] MOVED pandas dependent logic of get_expected_pred_idx from fhv2 to converter utility
